+++ options/locale/locale_en-US.ini
@@ -721,6 +721,7 @@ migrate_items_milestones = Milestones
 migrate_items_labels = Labels
 migrate_items_issues = Issues
 migrate_items_pullrequests = Pull Requests
+migrate_items_merge_requests = Merge Requests
 migrate_items_releases = Releases
 migrate_repo = Migrate Repository
 migrate.clone_address = Migrate / Clone From URL
@@ -730,11 +731,15 @@ migrate.permission_denied = You are not allowed to import local repositories.
 migrate.invalid_local_path = "The local path is invalid. It does not exist or is not a directory."
 migrate.failed = Migration failed: %v
 migrate.lfs_mirror_unsupported = Mirroring LFS objects is not supported - use 'git lfs fetch --all' and 'git lfs push --all' instead.
-migrate.migrate_items_options = Authentication is needed to migrate items from a service that supports them.
+migrate.migrate_items_options = Access Token is required to migrate additional items
 migrated_from = Migrated from <a href="%[1]s">%[2]s</a>
 migrated_from_fake = Migrated From %[1]s
+migrate.migrate = Migrate From %s
 migrate.migrating = Migrating from <b>%s</b> ...
 migrate.migrating_failed = Migrating from <b>%s</b> failed.
+migrate.github.description = Migrating data from Github.com or Github Enterprise.
+migrate.git.description = Migrating or Mirroring git data from Git services
+migrate.gitlab.description = Migrating data from GitLab.com or Self-Hosted gitlab server.
 
 mirror_from = mirror of
 forked_from = forked from
