+++ templates/user/dashboard/repolist.tmpl
@@ -25,14 +25,6 @@
 		<div v-show="tab === 'repos'" class="ui tab active list dashboard-repos">
 			<h4 class="ui top attached header">
 				{{.i18n.Tr "home.my_repos"}} <span class="ui grey label">${reposTotalCount}</span>
-				{{if or (not .ContextUser.IsOrganization) .IsOrganizationOwner}}
-				<div class="ui right">
-					<a class="poping up" :href="suburl + '/repo/create{{if .ContextUser.IsOrganization}}?org={{.ContextUser.ID}}{{end}}'" data-content="{{.i18n.Tr "new_repo"}}" data-variation="tiny inverted" data-position="left center">
-						<i class="plus icon"></i>
-						<span class="sr-only">{{.i18n.Tr "new_repo"}}</span>
-					</a>
-				</div>
-				{{end}}
 			</h4>
 			<div class="ui attached secondary segment repos-search">
 				<div class="ui fluid right action left icon input" :class="{loading: isLoading}">
