+++ web_src/js/features/migration.js
@@ -7,7 +7,6 @@ const $items = $('#migrate_items').find('.field');
 export default function initMigration() {
   checkAuth();
 
-  $service.on('change', checkAuth);
   $user.on('keyup', () => {checkItems(false)});
   $pass.on('keyup', () => {checkItems(false)});
   $token.on('keyup', () => {checkItems(true)});
@@ -23,19 +22,8 @@ export default function initMigration() {
 
 function checkAuth() {
   const serviceType = $service.val();
-  const tokenAuth = $(`#service-${serviceType}`).data('token');
 
-  if (tokenAuth) {
-    $user.parent().addClass('disabled');
-    $pass.parent().addClass('disabled');
-    $token.parent().removeClass('disabled');
-  } else {
-    $user.parent().removeClass('disabled');
-    $pass.parent().removeClass('disabled');
-    $token.parent().addClass('disabled');
-  }
-
-  checkItems(tokenAuth);
+  checkItems(serviceType !== 1);
 }
 
 function checkItems(tokenAuth) {
