+++ web_src/less/_repository.less
@@ -3200,3 +3200,7 @@ td.blob-excerpt {
 .select-project .item .svg {
   margin-right: .5rem;
 }
+
+.migrate .cards .card {
+  text-align: center;
+}
