+++ web_src/less/themes/theme-arc-green.less
@@ -1994,3 +1994,19 @@ footer .container .links > * {
   border: 1px solid rgba(121, 71, 66, .5) !important;
   border-bottom: none !important;
 }
+
+.migrate .cards .card {
+  text-align: center;
+}
+
+.migrate .cards .card .content a {
+  color: rgb(158, 158, 158) !important;
+}
+
+.migrate .cards .card .content a:hover {
+  color: rgb(255, 255, 255) !important;
+}
+
+.migrate .cards .card .content .description {
+  color: rgb(158, 158, 158);
+}
